pyproject.toml
[project]
name = "glassbox-operator"
version = "0.2.0"
description = "Agentic company-operator platform: tasks, runs, policy-gated tools, evidence packs, RAG store, and Google Sheets integration."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [{name = "Glassbox Operator"}]

dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.0",
  "SQLAlchemy>=2.0.0",
  "httpx>=0.27.0",
  "tenacity>=8.2.0",
  "python-json-logger>=2.0.7",
  "orjson>=3.9.15",
  "typer>=0.12.0",
  "python-dotenv>=1.0.1",
  "python-docx>=1.1.0",
  "reportlab>=4.1.0",
  "numpy>=1.26.0",
  "Jinja2>=3.1.4",
  "PyYAML>=6.0.2",
  "prometheus-client>=0.20.0",
  "apscheduler>=3.10.0",
  "structlog>=24.1.0",
  "aiosqlite>=0.20.0",
  "asyncpg>=0.29.0",
  "feedparser>=6.0.0",
  "beautifulsoup4>=4.12.0",
  "python-dateutil>=2.9.0",
  "redis>=5.0.0",
  "celery>=5.3.0",
  "vaderSentiment>=3.3.2",
]

[project.optional-dependencies]
openai = ["openai>=1.40.0"]
google = [
  "google-api-python-client>=2.120.0",
  "google-auth>=2.29.0",
  "google-auth-httplib2>=0.2.0",
]

[project.scripts]
gbo = "app.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"
